fix: improve add landing page test
