+++ tests/acceptance/tasks/ShopAdmin/Category/CreateLandingPage.ts
@@ -2,16 +2,10 @@ import { test as base } from '@playwright/test';
 import type { FixtureTypes, Task } from '@fixtures/AcceptanceTest';
 
 export const CreateLandingPage = base.extend<{ CreateLandingPage: Task }, FixtureTypes>({
-    CreateLandingPage: async ({ ShopAdmin, AdminCategories, AdminLandingPageCreate, AdminLandingPageDetail, IdProvider }, use ) => {
+    CreateLandingPage: async ({ ShopAdmin, AdminCategories, AdminLandingPageCreate, AdminLandingPageDetail }, use ) => {
 
-        const landingPageData = {
-            name: `Landing Page ${IdProvider.getIdPair().uuid}`,
-            salesChannel: 'Storefront',
-            seoUrl: `landing-${IdProvider.getIdPair().id}`,
-        };
-        const task = (layoutName: string, status: boolean) => {
+        const task = (layoutName: string, landingPageData, status: boolean) => {
             return async function CreateLandingPage() {
-
                 await AdminCategories.landingPageHeadline.click();
                 await AdminCategories.addLandingPageButton.click();
 
@@ -42,28 +36,6 @@ export const CreateLandingPage = base.extend<{ CreateLandingPage: Task }, Fixtur
                 }
                 await AdminLandingPageCreate.saveLandingPageButton.click();
                 await AdminLandingPageCreate.loadingSpinner.waitFor({ state: 'hidden' });
-
-                // Verify created landing page
-                const createdLandingPage = AdminCategories.landingPageItems.locator(`text="${landingPageData.name}"`);
-                await createdLandingPage.click();
-                await AdminLandingPageDetail.loadingSpinner.waitFor({ state: 'hidden' });
-
-                // Verify general tab detail
-                await ShopAdmin.expects(AdminLandingPageDetail.nameInput).toHaveValue(landingPageData.name);
-                await ShopAdmin.expects(AdminLandingPageDetail.landingPageStatus).toBeChecked();
-                await ShopAdmin.expects(AdminLandingPageDetail.salesChannelSelectionList).toHaveText(landingPageData.salesChannel);
-                await ShopAdmin.expects(AdminLandingPageDetail.seoUrlInput).toHaveValue(landingPageData.seoUrl);
-                // Verify layout tab detail
-                if (layoutName) {
-                    await AdminLandingPageDetail.layoutTab.click();
-                    await ShopAdmin.expects(AdminLandingPageDetail.layoutAssignmentCardTitle).toHaveText(layoutName);
-                    await ShopAdmin.expects(AdminLandingPageDetail.layoutAssignmentCardHeadline).toHaveText(layoutName);
-
-                    await ShopAdmin.expects(AdminLandingPageDetail.layoutAssignmentContentSection).toBeVisible();
-                    await ShopAdmin.expects(AdminLandingPageDetail.layoutResetButton).toBeVisible();
-                    await ShopAdmin.expects(AdminLandingPageDetail.changeLayoutButton).toBeVisible();
-                    await ShopAdmin.expects(AdminLandingPageDetail.editInDesignerButton).toBeVisible();
-                }
             }
         }
