+++ tests/acceptance/tasks/ShopAdminTasks.ts
@@ -20,9 +20,15 @@ import { FRWSalesChannelSelectionPossibility } from '@tasks/ShopAdmin/FRW/FRWSal
  */
 import { CreateLandingPage } from '@tasks/ShopAdmin/Category/CreateLandingPage';
 
+/**
+ * Verify Landing Page From Category
+ */
+import { VerifyLandingPage } from '@tasks/ShopAdmin/Category/VerifyLandingPage';
+
 export const test = mergeTests(
     GenerateVariants,
     UploadImage,
     FRWSalesChannelSelectionPossibility,
     CreateLandingPage,
+    VerifyLandingPage,
 );
